# src/main/resources/static/api/openapi.yml
openapi: 3.1.0
info:
  title: BOS API documentation
  description: |
    내꿈은건물주  API 문서

  version: 0.0.1-SNAPSHOT
  contact: {}
...
paths:
...
  /push/test/simple:
    post:
      tags:
        - push
      operationId: testSimplePush
      summary: "간편 푸시 테스트 - userId만으로 자동 테스트"
      description: |-
        userId만 입력하면 자동으로 최신 거래 내역과 다음 상환 스케줄을 조회하여 푸시 전송

        ## 자동 조회 항목
        - 최신 거래 내역 (createdAt 기준)
        - 다음 상환 스케줄 (첫 번째 pending 스케줄)
        - 등록된 모든 디바이스

        ## 발송 템플릿
        - LEND (빌려준 경우): REPAYMENT_REMINDER_PAYEE
        - BORROW (빌린 경우): REPAYMENT_REMINDER_PAYER

        ## 응답 정보
        - 푸시 전송 결과와 함께 사용된 데이터 정보를 반환
        - 테스트에 사용된 거래, 스케줄, 템플릿 정보 포함
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SimplePushTestRequest"
      responses:
        "200":
          description: 푸시 전송 완료
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimplePushTestResponse"
        "400":
          description: 잘못된 요청
        "500":
          $ref: "#/components/responses/InternalServerErrorResponseSpec"
  /push/test/repayment-reminder-payer:
    post:
      tags:
        - push
      operationId: testRepaymentReminderPayer
      summary: "상환 예정 안내 (갚을 사람) 테스트"
      description: |-
        상환 예정일 D-2 오전 9시에 발송되는 푸시 알림 템플릿 테스트

        ## 발송 조건
        - 상환일 2일 전 오전 9시 발송
        - 예정된 회차 존재
        - 상환 완료 시 발송 안 됨
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RepaymentPushTestRequest"
      responses:
        "200":
          description: 푸시 전송 완료
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PushTestResponse"
        "400":
          description: 잘못된 요청
        "500":
          $ref: "#/components/responses/InternalServerErrorResponseSpec"
  /push/test/repayment-reminder-payee:
    post:
      tags:
        - push
      operationId: testRepaymentReminderPayee
      summary: "상환 예정 안내 (받을 사람) 테스트"
      description: |-
        상환 예정일 D-2 오전 9시에 발송되는 푸시 알림 템플릿 테스트

        ## 발송 조건
        - 상환일 2일 전 오전 9시 발송
        - 예정된 회차 존재
        - 상환 완료 시 발송 안 됨
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RepaymentPushTestRequest"
      responses:
        "200":
          description: 푸시 전송 완료
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PushTestResponse"
        "400":
          description: 잘못된 요청
        "500":
          $ref: "#/components/responses/InternalServerErrorResponseSpec"
  /push/test/repayment-today-payer:
    post:
      tags:
        - push
      operationId: testRepaymentTodayPayer
      summary: "상환일 당일 안내 (갚을 사람) 테스트"
      description: |-
        상환일 당일 오전 9시에 발송되는 푸시 알림 템플릿 테스트

        ## 발송 조건
        - 상환일 당일 오전 9시 발송
        - 예정된 회차 존재 + 해당 회차 미완료
        - 당일 0시~발송 시점 사이 상환 완료 시 발송 안 됨
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RepaymentPushTestRequest"
      responses:
        "200":
          description: 푸시 전송 완료
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PushTestResponse"
        "400":
          description: 잘못된 요청
        "500":
          $ref: "#/components/responses/InternalServerErrorResponseSpec"
  /push/test/repayment-today-payee:
    post:
      tags:
        - push
      operationId: testRepaymentTodayPayee
      summary: "상환일 당일 안내 (받을 사람) 테스트"
      description: |-
        상환일 당일 오전 9시에 발송되는 푸시 알림 템플릿 테스트

        ## 발송 조건
        - 상환일 당일 오전 9시 발송
        - 예정된 회차 존재 + 해당 회차 미완료
        - 당일 0시~발송 시점 사이 상환 완료 시 발송 안 됨
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RepaymentPushTestRequest"
      responses:
        "200":
          description: 푸시 전송 완료
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PushTestResponse"
        "400":
          description: 잘못된 요청
        "500":
          $ref: "#/components/responses/InternalServerErrorResponseSpec"
  /push/test/repayment-overdue-payer:
    post:
      tags:
        - push
      operationId: testRepaymentOverduePayer
      summary: "상환 지연 경고 (갚을 사람) 테스트"
      description: |-
        상환일 익일 오전 9시에 발송되는 푸시 알림 템플릿 테스트

        ## 발송 조건
        - 상환일 익일 오전 9시 발송
        - 예정된 회차 존재 + 해당 회차 미완료
        - 상환일 당일 23:59까지 완료 시 발송 안 됨
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RepaymentOverduePushTestRequest"
      responses:
        "200":
          description: 푸시 전송 완료
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PushTestResponse"
        "400":
          description: 잘못된 요청
        "500":
          $ref: "#/components/responses/InternalServerErrorResponseSpec"
  /push/test/repayment-overdue-payee:
    post:
      tags:
        - push
      operationId: testRepaymentOverduePayee
      summary: "상환 지연 경고 (받을 사람) 테스트"
      description: |-
        상환일 익일 오전 9시에 발송되는 푸시 알림 템플릿 테스트

        ## 발송 조건
        - 상환일 익일 오전 9시 발송
        - 예정된 회차 존재 + 해당 회차 미완료
        - 상환일 당일 23:59까지 완료 시 발송 안 됨
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RepaymentOverduePushTestRequest"
      responses:
        "200":
          description: 푸시 전송 완료
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PushTestResponse"
        "400":
          description: 잘못된 요청
        "500":
          $ref: "#/components/responses/InternalServerErrorResponseSpec"
  /push/test/partial-repayment-complete:
    post:
      tags:
        - push
      operationId: testPartialRepaymentComplete
      summary: "부분 상환 완료 안내 테스트"
      description: |-
        상환 1회 완료 시 발송되는 푸시 알림 템플릿 테스트

        ## 발송 조건
        - 상환 1회 완료 시 즉시 발송
        - 진행률 N% → N% 변경
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PartialRepaymentCompletePushTestRequest"
      responses:
        "200":
          description: 푸시 전송 완료
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PushTestResponse"
        "400":
          description: 잘못된 요청
        "500":
          $ref: "#/components/responses/InternalServerErrorResponseSpec"
  /push/test/transaction-complete:
    post:
      tags:
        - push
      operationId: testTransactionComplete
      summary: "거래 내역 최종 완료 안내 테스트"
      description: |-
        해당 거래 내역 100% 완료 시 발송되는 푸시 알림 템플릿 테스트

        ## 발송 조건
        - 전체 금액 상환 완료 시 즉시 발송
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionCompletePushTestRequest"
      responses:
        "200":
          description: 푸시 전송 완료
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PushTestResponse"
        "400":
          description: 잘못된 요청
        "500":
          $ref: "#/components/responses/InternalServerErrorResponseSpec"
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    InternalServerErrorResponseSpec:
      description: 내부 오류가 생겼습니다. 해당 상황 발생 시, 관리자에게 이야기해주세요.
      content:
        application/json:
          schema:
            properties:
              statusCode:
                type: number
                example: 500
              message:
                type: string
                example: Internal Server Error
...